Move freezables into separate files.
