+++ polynomial/__init__.py
@@ -3,13 +3,11 @@
 (c) Yalishanda <yalishanda@abv.bg>
 """
 
-from .frozen import Freezable
 from .core import (
     Constant,
-    FrozenPolynomial,
     Monomial,
     Polynomial,
-    ZeroPolynomial
 )
+from .freezables import FrozenPolynomial, ZeroPolynomial
 from .binomial import Binomial, LinearBinomial
 from .trinomial import Trinomial, QuadraticTrinomial
